+++ packages/cli/src/utils/firebase.ts
@@ -1,0 +1,12 @@
+import Firebase from 'firebase'
+
+export default Firebase.initializeApp({
+  apiKey: "AIzaSyDp-sk6i2gICPGv66O13l_ttoitXu9849w",
+  authDomain: "platform-f4e92.firebaseapp.com",
+  databaseURL: "https://platform-f4e92.firebaseio.com",
+  projectId: "platform-f4e92",
+  storageBucket: "platform-f4e92.appspot.com",
+  messagingSenderId: "581848581612",
+  appId: "1:581848581612:web:44440228033335dbc52cc9",
+  measurementId: "G-VQKX4K7PTT"
+})