+++ packages/api/CHANGELOG.md
@@ -6,7 +6,7 @@
 #### Improvements
 #### Bug fixes
 
-- [#](https://github.com/mesg-foundation/js-sdk/pull/) Add missing functions `exists`/`hash` for `process` and `runner`
+- [#233](https://github.com/mesg-foundation/js-sdk/pull/233) Add missing functions `exists` and `hash` for `process` and `runner`
 
 ## [v0.3.0](https://github.com/mesg-foundation/js-sdk/releases/tag/%40mesg%2Fapi%400.3.0)