+++ packages/api/CHANGELOG.md
@@ -6,6 +6,10 @@
 #### Improvements
 #### Bug fixes
 
+## [v0.3.1](https://github.com/mesg-foundation/js-sdk/releases/tag/%40mesg%2Fapi%400.3.1)
+
+#### Bug fixes
+
 - [#233](https://github.com/mesg-foundation/js-sdk/pull/233) Add missing functions `exists` and `hash` for `process` and `runner`
 
 ## [v0.3.0](https://github.com/mesg-foundation/js-sdk/releases/tag/%40mesg%2Fapi%400.3.0)